fix csv writer
